Reorganize submission tests
We currently only have a handful of submission tests in the modernized portion of the test suite.  As more of our legacy tests get converted, the current haphazard organization will become more of a burden.  This PR organizes the tests by submission type, with each directory having a child data directory.  Most new test cases should be added to the existing test files.  In rare cases which require special handling (a test which needs RUN_SERIAL, for example), an additional test file in the same directory may be created in the future.
